sistema

+++ dzsistema/src/Negocio/Entidades/FacturaJpaController.java
@@ -134,7 +134,7 @@ public class FacturaJpaController {
         }
     }
 
-    public Factura buscarPorNumero(int numero) {
+    public Factura buscarPorNumero(long numero) {
         EntityManager em = getEntityManager();
         List<Factura> encontrados;
         encontrados = em.createQuery("SELECT a FROM Factura a WHERE a.numero = '"+numero+"'").getResultList();
